[perf][e2e] Support richer metrics explanations in metrics sets
It's common for test authors who capture a series of measurements during
a test to use trace_utils.standard_metrics_set to calculate averages and
percentiles. Before this patch, this function returned a set of TestCaseResult
objects that just re-used the metric name as a short metric description.
This patch allows callers to provide a richer descriptive string, defaulting
to the prior behavior.

Bug: b/383590353
Test: trace_utils_test

Change-Id: I04729a4592b1f223db3ad2412349be844e9aab66
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1289468
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Commit-Queue: Chris Masone <cmasone@google.com>