[starnix][vfs] Remove FsNodeInfo::ino
This field is redundant with `FsNode::node_id`. This CL also renames
`FsNode::node_id` to `FsNode::ino` to be clearer about what this field
means.

In order to implement FUSE correctly, we need to have a separate nodeid
from the ino for FUSE nodes. However, FUSE already has storage for that
value in the FuseNode struct. This CL adds an FsNodeOps method so that
we can reuse this storage rather than having an extra field on `FsNode`
that is almost always the same as the ino.

Bug: b/419306849
Change-Id: Ic530e4b431b799f33108d8d93300eb59d83e68d4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1289246
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>
Reviewed-by: James Robinson <jamesr@google.com>
Commit-Queue: Adam Barth <abarth@google.com>