[audio][tests] Deflake audio_driver_admin_tests
Add a delay after unbinding signalprocessing, to make these
test cases more reliable. The intermittent failure is seen
as ZX_ERR_ALREADY_BOUND, which indicates that we are not
waiting long enough, after disconnecting from the driver,
before reestablishing a FIDL connection.

We don't have an ironclad way to know with certainty how
long a "cooldown" delay we need, or necessarily even which
interface is causing the problem. However, because all the
recent work on audio_driver_tests has been in the area of
signalprocessing tests, the logical conclusion is that the
cooldown should be added to the signalprocessing protocol.

We'll start by inserting a modest delay of 10 ms, and see
if the ALREADY_BOUND failures are eliminated.

Fixed: 416759183
Multiply: audio_driver_admin_tests

Change-Id: I60af6bfd07f72e29fbe81429c1ef23fac8a05954
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1279375
Fuchsia-Auto-Submit: Martin Puryear <mpuryear@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>