[time][alarms] Remove the overly strict assertion
Replaces a check which ensures the "now" is always less than deadline
with a limit that replaces all negative timeouts with a zero.

Also required changes in the code that assumed a deadline of zero would
not happen. Guarded with a unit test.

Bug: 421507068
Change-Id: I561cc6efb317bdd953d52a6c1a873cb38609c574
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1289026
Reviewed-by: Francois Rousseau <frousseau@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>