[bazel] Fix go toolchain module definition for MacOS.
Introduce the @fuchsia_sdk_common module to wrap
//build/bazel_sdk/bazel_rules_fuchsia/common properly
without Bazel errors.

Ensure @fuchsia_build_config//:host_prebuilts/ contains
symlinks to various host prebuilt directories and tools
without an explicit Fuchsia host-tag in their path,
then use them in a few module and repository definitions.

This fixes the Go toolchain wrapper which can now work
properly on MacOS.

Note that it is not possible to use these symlinks
for the python, clang and rust repositories for now
because their repository rule require a path string
instead of a label to locate the corresponding
directories or tools. This will be fixed in a future
CL that will modify these rules to accept a string
that accepts either a path or a Bazel label as
argument.

Fixed: 421218053
Change-Id: Icb164f0b8d8430723427b5529ca8a6caccd68903
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1289604
Reviewed-by: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: David Turner <digit@google.com>