[kernel][vm] Avoid cur_locked_ holding incorrect value
cur_locked_ should only hold a locked version of cur_, and never
anything else. UpdateCurLocked presently violates this as it uses
cur_locked_ while attempting to get a synchronized version of a lock
to cur_. This can cause an assert in MoveCurLocked to spuriously trip
as it assumes that since it holds cur_->lock(), that cur_locked_ cannot
have a value.

The fix in this commit is to use a stack temporary cur_locked until it
is known that the current lock was acquired, then move to cur_locked_.

Bug: 421025570
Change-Id: Ib5efb46e4172a5f4dd6f1f519ed9d166f9da7f7e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1288044
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>