[sestarnix] Test remapping of file descriptors on exec
When a domain transition occurs on exec and the
post-exec domain does not have the 'use' permission
for a file descriptor opened in the pre-exec context,
then the file descriptor is remapped to the 'null'
file in selinuxfs. This change adds tests for this
behavior.

Bug: 322843830

Change-Id: I507b83722163ce4d073298c05371008876e1ab83
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1257993
Reviewed-by: Wez <wez@google.com>
Commit-Queue: Laura Peskin <pesk@google.com>