[debug][inspector] Avoid test timeout flake
In a nested virtualization environment it is possible for the
loop_thread_'s to not be blocked before they are suspended as part of
the exception report. If this happens the threads suspension may happen
asynchronously, and could spuriously timeout.

This change waits for the test threads to enter the blocked state,
ensuring that suspension will be synchronous and hence can never
timeout.

Bug: 420598541
Change-Id: I8802f9c3dba4a84ba3ca8e58ffd164dd804c3289
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1287904
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>