[netlink] Use a high metric for the backup route
When a route is added/removed through netlink, we create a copy of it
in the main table, we will use the priority attribute of the netlink
message as the metric for the route. This could lead to accidentally
overriding the metric we assign to the interfaces. This change fixes
the problem by applying a delta on the metric to deprioritize the
copied route in the main table. More details in the attached bug.

Fixed: 418059666
Change-Id: I6a6a696490d9229c3fce0154951ff260329c9452
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1278645
Commit-Queue: Zeling Feng <zeling@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>