[graphics][magma] Add magma_enumerate_devices
This is a convenience API so clients don't have to implement
listing directory contents.  The API is able to open devices
from the current devfs-based /loader-gpu-devices/class/gpu or
the new services-based /loader-gpu-devices/svc/magma. The latter
is recommended by the define MAGMA_DEVICE_NAMESPACE.

If an ICD wishes to be backward compatible it can call this
on MAGMA_DEVICE_NAMESPACE and if it finds nothing it can call
it with the devfs path instead.

Bug: 419087951
Change-Id: I420e17f3146a42efc02f2d1a8dde7649a6ac972f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1283465
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Fuchsia-Auto-Submit: Craig Stout <cstout@google.com>
API-Review: Victor Costan <costan@google.com>
Reviewed-by: Victor Costan <costan@google.com>
Commit-Queue: Craig Stout <cstout@google.com>