[rust-3p] cargo-vendor now includes a number of additional files
We rolled the rust toolchain on Friday, which contains an update to
cargo https://github.com/rust-lang/cargo/pull/15514, which migrates from
using a heuristic to select the files to include to instead use the
vendor/ directory. This should match the files that are in the .crate
archive. This now includes a number of additional files to vendor, but
otherwise should have no semantic changes.

Note that this updates predicates 1.0.7 => 1.0.8, which just removes a
large `.tags` file.

Change-Id: I95b638da4cd125a6c42c18edcb347771207f1ef3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1290710
Reviewed-by: David Koloski <dkoloski@google.com>
Reviewed-by: Thomas Zander <thomasalpinus@google.com>
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: Thomas Zander <thomasalpinus@google.com>