[fshost] fail all integration tests on error logs
Remove the escape hatch that allowed the storage-host fshost integration
tests to produce error logs without failing. This escape hatch allows
for regressions, bugs, and flakes to be merged and fester that would
otherwise be caught immediately.

I think there were other errors a while ago but at this time there were
only two - for the tests where we initialize the gpt at runtime, the gpt
component printed an error. There is probably a better way to fix it,
but as the comment in the same area says, returning errors isn't
currently something we can do. Second was the gpt component was
sometimes getting shut down twice, when it wasn't being used. This is
fixed by making double shutdown no longer be an error case - it seems
fine to just say we are already in the state we need and leave it at
that.

Fixed: 391891190
Test: fx test fshost_integration_tests
Change-Id: Idd64aefe884fd334a7bc0e1e9c92dfdcbc6c545e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1290647
Commit-Queue: Stephen Demos <sdemos@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>