[trace] Making sure ATrace timestamps always increase
This moves the timestamp collection for Atrace events to be inside the
TraceEventQueue struct which  protects the event queue metadata with
a RWLock.

The result is that the timestamp for each event written, regardless of
which thread is writing, is greater than the previous timestamp.

The effect of this change is that "End" events are always written
aftert the "Begin" events.

Bug: 401645216
Change-Id: I5954061bfa88994ba9990592b05b84a5dde2e467
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1288491
Reviewed-by: Jia Li <lizhijia@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: David Song <wintermelons@google.com>
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>