[starnix] Remove unneeded kernel mapping for FUTEX_(UN)LOCK_PI.
Now that usercopy can support atomic compare/exchange we can use that
instead of making a mapping for ourselves on every lock contention.

In addition to removing the now-obsolete mapping, this also changes
the futex table operations to always use acquire/release ordering.
The previously-used relaxed ordering does not create synchronization
points with other CPUs which is needed for this code to be correct
when userspace only does the fast-path in a mutex without making a
syscall. See bionic's atomic orderings in mutex[0] for an example:
https://cs.android.com/android/platform/superproject/main/+/main:bionic/libc/bionic/pthread_mutex.cpp;l=148-231;drc=2e65afecba6877eaffa6bda825ae156e6c682fe5

Change-Id: Ibd387289b7c563e731b4345fd76f95b2c7a9500e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1247369
Reviewed-by: James Robinson <jamesr@google.com>
Reviewed-by: Theodore Dubois <tbodt@google.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>
Commit-Queue: Adam Perry <adamperry@google.com>