[kernel][vm] Explicit VmPageSpliceList life cycle
The VmPageSpliceList is a large (~300 byte) object, and any
construction or move assignment has a non-trivial cost. This changes
the way the VmPageSpliceList can be initialized so that it can be
passed around as a pointer instead of needing to build new ones and
perform a move assignment.

This change is done by having more explicit life cycle management,
allowing certain initialization steps to be performed separately to
construction.

Change-Id: I8c238b3054f6a69ac1f3c0c7389d12356b72d3f8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1286984
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>