[fshost] downgrade watcher stream error to warn
Sometimes when shutting down the gpt component, the watcher for it gets
interrupted with PEER_CLOSED. It seems to only happen on
write_data_file, but it's not clear why. By downgrading to a warning for
now, we can prevent the flakes until we figure out a root cause.

Bug: 422230360
Change-Id: I5522fa5c69c29f513453081c17b6ef0a08bbec07
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1291432
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>