[assembly] Introduce emulator_support_aib.gni
Initially the custom boot shim support was introduced by querying the
target's name and making an assumption on the output.

While it is possible to change all callsites to match this convention,
there are tools downstream (ffx emu) that make decisions on the filename,
such as whether to add some extra arguments to QEMU based on .efi
extension.

To remove this inaccurate restriction and allow the unification of qemu
kernel tests and boot tests, we provide a metadata driven mechanism to
obtain the path from a qemu kernel.

phys_executable and linuxbiboot_executable now have an additional
parameter to optionally produce `emulator_support_aib_input` data and
walk key (barrier version of the data key). Scripts are introduced to
generate an assembly input bundle for the shim.

From this point on it hooks into existing mechanisms to override the
default qemu kernel.

`shim_executable` is now forwarded any `qemu.import` such that certain
templates are imported for defining the proper targets.

Bug: 418057728
Change-Id: I2a993554933a6c9941206ebc459672d4c4c06129
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1278685
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>