[boot_test] Use a default empty non-bootable ZBI
This allows using the same code path for all boot tests, whether they
are emulator only or not.

`efi_test` has been slightly modified so it fits with the executables
generated by other similar tests that generate a qemu kernel.

Next step is to work with assembly and define some metadata to find the
path to the qemu kernel and not make any assumptions on the path of such
things.

Abuse bootloader-partition::type to indicate that the image is in fact a
EFI Shell script. This now removes the need to require no system_a to
identify this.

In a similar fashion, ffx emu was only looking at this partitions when
system_a was missing. Now they check if a single bootloader partition of
type "efi-shell" is present.

Bug: 414683197
Change-Id: I2e0979a00e321dfcd680290a55674b4339366a7c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1277291
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>
Reviewed-by: Yilong Li <liyl@google.com>