[assembly] Move PB creation to a crate
Introduces ProductBundleBuilder that can be used to programmatically
create a product bundle and write it to a directory. This will be
used in 'ffx product' as well as bisection tooling.

The old logic is moved to a separate function and the output is
diffed against the output of ProductBundleBuilder. A follow-up CL
will delete the old logic.

Bug: 421192283
Change-Id: I0d6767ddd65dab7090a618ee88e913da1a2d62f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1289386
Reviewed-by: Jerry Belton <jcecil@google.com>
Commit-Queue: Aidan Wolter <awolter@google.com>