[bzlmod] Migrate @re2 to bzlmod
re2 is required for newer version of googletest, which is later needed
to migrate @fuchsia_sdk.

@abseil-cpp is moved as well because it's a dependency.

NOTE: We already have a version of re2 at //third_party/re2, so OSRB
review is not necessary. We need a newer version for bzlmod and the
existing one is hard to uprev, so we vendor a newer version as a
stopgap. See linked bug for details.

Bug: 421339820

Change-Id: I41df3ea8b42570d300514d51a1184346d3cd2e25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1291024
Commit-Queue: Jay Zhuang <jayzhuang@google.com>
Reviewed-by: Jerry Belton <jcecil@google.com>
Reviewed-by: David Turner <digit@google.com>