[starnix] ignore data before read/write_consumed in BINDER_READ_WRITE
The binder_write_read structure has a write_consumed and read_consumed
field that are filled by the kernel. When these fields are non-zero
before the ioctl call, the first write_consumed/read_consumed bytes
of the buffers should be ignored. This allows, e.g., restarting the
call with the same structure when getting EINTR.
The value should also be updated with each command processed, so that
in case of error it indicates what commands where fully processed.

This behavior is relied upon by the SELinux binder tests.

Bug: 420388565
Change-Id: I2882738e375bb0941c752932a33c666d9d2bfda0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1289610
Reviewed-by: Benjamin Lerman <qsr@google.com>
Commit-Queue: Ambre Williams <ambre@google.com>