[logs] Assert that the logging pipeline is non-waiting
Modify the serial logging test to ensure
that writes are immediately visible to readers.
This guarantee is useful in implementing a "flush"
mechanism in Archivist for logs. Our "flush" mechanism
can simply wait for all ingestion futures to return
Poll::Pending, which will indicate that we've processed
all logs in the pipeline since before the Flush
call was invoked. Previously we had no test which
asserted on this property.

Bug: 386831734
Change-Id: Ic5b9d38f18e21ece6345cf068cd8512c2fbdeb00
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1292524
Reviewed-by: Clayton McCray <claytonmccray@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>