[testing][time] A high fidelity fake hrtimer service
This is a fake service which implements the driver API service
`fuchsia.hardware.hrtimer/Service`. It is intended to be used
in hermetic e2e tests which may not access the "system" wake
alarms device.

This is a rust reimplementation of the C++ fake hrtimer which was
originally created for purposes of power broker tests, but has since
outrgown its original purpose.

The reason the service is reimplemented is because the C++ fake hrtimer
has limited fidelity specifically regarding handling simultaneous client
requests, leading to effective serialization of alarms. This is
functionally correct, but leads to long delays when making multiple
requests, leading to false detection of performance regressions, and a
higher test runtime variance when compared to non-hermetic tests.

The C++ fake hrtimer uses the synchronous FIDL API which means adjusting
its fidelity for concurrency support would be a rewrite anyway. So when
rewriting I opted for a rust reimplementation, which is the same code
base that all its now-companion binaries are written in.

The intention is to merge this new component, then gradually replace the
old fake hrtimer service in the tests it is used in. Once that is done,
the original C++ fake hrtimer service can be deleted.

Tested: locally with relevant e2e test runs
Bug: 421294594
Change-Id: Ie920d5c2196d0b5d867da099f44ab237e0140d03
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1291968
Reviewed-by: Guocheng Wei <guochengwei@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>