[dl] Fill out direct_deps of startup modules
A second-pass loop is added to
RuntimeDynamicLinker::PopulateStartupModules to set the direct
dependencies of each startup module that was initialized by the
previous loop in the function.

To allow for fallible allocation without a Diagnostics object, parent
class functions of the AllocCheckerContainer (ie fbl::Vector) are made
public, and `push_back()` is used to fill out direct_deps in the
second-pass loop.

The multiple-transitive-foo-deps module is converted into a
`test_shared_library` so that it may be linked in as a dependency for
dl-unittests (ie.e dl-unittest.deps) to use as a startup module for
testing. To make sure this target is linked for the existing test user
(LoadedTransitiveDepOrder), a `per_test_shlibs` array is introduced to
import standalone test-shared-libraries.

Fixed: 379766260
Change-Id: I3e69c445a93d7b52d1e506086ddfa1cd2e3bda58
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1290707
Commit-Queue: Caslyn Tonelli <caslyn@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Fuchsia-Auto-Submit: Caslyn Tonelli <caslyn@google.com>