[rustc_test] Support backtraces from host tests.
When running on the host, we currently do not get backtraces from test
failures for two reasons:

1) We need to set RUST_BACKTRACE=1 in the environment.
2) We run stripped binaries which lack the information to symbolize the
   stack.

This change passes the above environment variable and points test
clients at the unstripped binary to support backtraces. To ensure
that tests can still access their host test data, the unstripped
binary is copied to the top level host output directory.

To avoid changing a large number of test names (which are used for
grouping, analysis, and test selection), this CL also adds a "name"
parameter to host_test which supports overriding the name in tests.json.
We then use this feature to preserve the original names.

Fixed: 421917078
Run-All-Tests: True
Change-Id: I841ebd554d43c1a2cadead726011f346d7fed1a5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1290564
Reviewed-by: David Fang <fangism@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Christopher Johnson <crjohns@google.com>