[starnix] Use DowncastedFile<SocketFile> as the type for check_socket_connect_access
Checking socket access needs to be done with a socket that has been
embodied in an FsNode. We should encode this in the type system.

The plan is to write future CLs to gradually make similar changes for
the rest of the socket access check hooks.

Bug: 414583985
Change-Id: I4bd664c347fda1401f924dea27bc4ad3b97ca802
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1286308
Reviewed-by: Wez <wez@google.com>
Commit-Queue: Theodore Dubois <tbodt@google.com>