[ffx] Do not err on empty nodename fastboot proxies.
This is useful for cases in which we simply want to identify a device by
its IP address and we do not have access to any other information about
a device. This can be used to identify unknown devices, allowing for
them to be merged with devices whose serial numbers are known.

Change-Id: Ie4a93f879ee7617c433d4483953c85306a6a636c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1287650
Fuchsia-Auto-Submit: Andrew Davies <awdavies@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>