[flatland] Warn when CpuRenderer generates incorrect images.
CpuRenderer cannot properly render all of the scenes supported by the
Flatland API.  Log a warning when an unsupported case is detected.
Add references to bugs in the code.

UNFORTUNATELY, this is pretty useless because we log DEBUG instead
of WARNING, so this probably won't help anyone who falls into this
pitfall, because nobody turns on DEBUG.  But we cant use WARNING
because log-spam because reasons.

Bug: 422817637
Bug: 406902418
Bug: 406902798
Bug: 406903965
Change-Id: I7efdbc8b2797a2a247e3e53bd27d482c06cff0cc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1238729
Reviewed-by: Jianpeng Chao <chaopeng@google.com>
Reviewed-by: Caroline Liu <carolineliu@google.com>
Commit-Queue: Josh Gargus <jjosh@google.com>