[clock][core] Add a node_id property to each clock node.
Previously if there was a duplicate Clock ID in the clock_ids metadata,
the core clock driver would fail to start because it would try to
publish two separate nodes with identical service names.

This condition would occur when two device tree nodes referenced the
same clock. The device tree parser would find the node twice and add it
to the metadata twice.

In this CL we additionally assign a node_id to each clock node which is
globally unique per clock controller. The node_id is also added as a
bind property which is used as a bind rule target for the composite node
spec created by the visitor.

Test: fx test clock-test
Bug: b/421487853
Change-Id: I81a53d5e0e2ffb834f5ce7a89fcfb142d8c8078b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1289668
Reviewed-by: Novin Changizi <novinc@google.com>
API-Review: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Gurjant Kalsi <gkalsi@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>