[starnix] u32 -> u8 for from_binder parameters
from_binder's one call-site creates the values it passes for these
parameters by masking and shifting such that they can only have set
bits in their lowest byte, so we can simply accept one byte for each
of them.

We also rename the "priority" parameter to "priority_or_niceness",
since we're going to be making a stronger distinction between the two,
and we rename several fields in the implementation for clarity.

This simplifies a following larger change to SchedulerPolicy.

Bug: 397371605
Change-Id: I6f71b8498547bf826fd8b39df87df7558c9b0b24
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1150334
Commit-Queue: Nathaniel Manista <nathaniel@google.com>
Reviewed-by: Benjamin Lerman <qsr@google.com>