[wlan] Fix target security bug in RoamRequestTest
This change fixes a bug in RoamRequestTest where hostapd would fail to
start for the target AP. This occurred when the origin AP security type
was open and the target AP security type was not. The target AP would
fail to start, as it was given an empty credential.

This change ensures a target credential is created when required, even
if the origin AP does not require one.

Change-Id: Id9f1235db83e5cbd5fa08ce47de9c337459f2e3e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1291438
Fuchsia-Auto-Submit: Hayden Nix <haydennix@google.com>
Reviewed-by: Karl Ward <karlward@google.com>
Commit-Queue: Hayden Nix <haydennix@google.com>