[unwinder] Add ArmEhAbi Frame Trust
This is just introducing the new Frame Trust type for users of this
library to handle if they require it. There are no behavior changes yet
and this is not expected to be used until it is enabled in a followup.

Bug: 395606674
Change-Id: I40ca1e288d88a533e4f4e27390fd4aa37d10847f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1272184
Commit-Queue: Jacob Rutherford <jruthe@google.com>
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Reviewed-by: Adam Barth <abarth@google.com>