[bazel][idk] Add idk_cc_source_library()
Add an initial implementation of `idk_cc_source_library()`, which is the
equivalent of the GN template `sdk_source_set()`. Also add an initial
implementation of `idk_atom()`, which is the equivalent of the GN
template `sdk_atom()`.

Define `FuchsiaIdkAtomInfo`, which describes an atom, and add files
demonstrating use of cquery and aspects with it.

Add a Bazel implementation of `memory_barriers` using the new
functionality.

Bug: 417305295, 417304469
Test: fx bazel build //build/bazel/bazel_idk/tests:build_only_tests && fx bazel build build/bazel/bazel_idk/tests:test-source-set_idk "--aspects=build/bazel/bazel_idk/idk_atom_info.bzl%print_deps_aspect" && fx bazel cquery build/bazel/bazel_idk/tests:test-source-set_idk --output=starlark --starlark:file=build/bazel/bazel_idk/idk_atom_info.cquery && fx bazel cquery build/bazel/bazel_idk/tests:test_molecule --output=starlark --starlark:file=build/bazel/bazel_idk/idk_molecule_info.cquery
Change-Id: I2e94c32204beb728c5c1b4559263d16206d1fbb3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1273895
Commit-Queue: David Dorwin <ddorwin@google.com>
Reviewed-by: David Turner <digit@google.com>
Fuchsia-Auto-Submit: David Dorwin <ddorwin@google.com>