[ffx] move ffx-driver from `plugins` to `tools`
In general, an ffx "plugin" is embedded in the binary, whereas a tool is a separate binary (in this case "ffx-driver"). We'd like to move more plugins to be tools, for clarity and for build-time reasons.

Recently we removed the dead code that supported ffx-driver being embedded in ffx, so this move reflects the reality that it is now a full-fledged tool.

Change-Id: I6a6a6964b00a6b68a3766bf044c7ccfca4b87f0d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1286768
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: Novin Changizi <novinc@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>