[fshost] Use in-process server for zbi ramdisk
When storage-host is enabled, fshost will locally host a block server
instance with the ramdisk ZBI's contents, rather than spawning a
separate driver.  We don't need any of the extra features that the
ramdisk driver offers, and this is simpler.

The block server runs in a dedicated thread, which avoids deadlocks if
fshost is single-threaded (fshost sends block requests to the ramdisk
server).

Fixed: 422781837
Change-Id: I1b1e6a30397b7bc5073405657b0b37479ce0d29e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1293845
Reviewed-by: Stephen Demos <sdemos@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>