[perfetto_consumer] Better handle startup errors
One bots, we'd sometimes see:
```
10:36:58.239224 [00029.126707][kernels:L7kol88][kthread] ERROR:
   [src/starnix/modules/perfetto_consumer/lib.rs(345)] perfetto_consumer
   callback error: ECONNREFUSED(111) from
   ../../src/starnix/kernel/vfs/socket/socket_unix.rs:1023:21
```
in the logs. This is from us getting notified that a trace is starting,
but the starnix perfetto consumer failed to connect to perfetto. Likely
we're too early in boot and perfetto hasn't actually started yet.

Unfortunately, we don't update our locally kept state unless we
successfully handle the
update.

Our use of short circuiting `?;` means we never get there. Once we
desync, the Stop/Stopping code checks that the previous state was
started, but it isn't, and we never release the prolonged context we
acquired.

We can fix this by always updating the local state regardless of if we
fail.

Bug: 411727604
Change-Id: I8c1d01899d5423524065fe01e49ba62acfb53fba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1295087
Fuchsia-Auto-Submit: Gwen Mittertreiner <gmtr@google.com>
Reviewed-by: Jia Li <lizhijia@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>