[kernel][suspend] Add a DSB before WFI in ARM Idle
According to section K14.2.3 of the ARM ARM (DDI 0487K.a), a DSB is
required before WFI/WFE if you want to be certain that all writes to
RAM are visible to other observers.

Add an explicit DSB to the ARM64 arch specific EnterIdleState to be
absolutely certain that all other PEs see all of the writes performed
by the PE entering idle before we do so.

Bug: 347699064
Change-Id: I29995f2cb9a492ffeddb6dca66925e53e3823bd0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1291442
Fuchsia-Auto-Submit: John Grossman <johngro@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: John Grossman <johngro@google.com>