[dhcp-client] Wait for Addr to become assigned
The DHCP client now uses the
AddressStateProvider::WatchAddressAssignmentState method to observe the
address transitioning from "Tentative" to "Assigned". The client waits
for up to the lease time to observe the address become assigned,
otherwise, it restarts.

As part of this change, refactor the way the core state machine is
notified of address removal/rejection. Originally this was done out-of
band using the `on_address_rejected` method. Now, the state machine
polls for address events in the main `run` function.

Bug: 42077260
Fixed: 420987904
Change-Id: I244d78d05e166859906f0f54789075b965dcb4a9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1286804
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>