[build][idk] Add prebuild info for atoms in all categories
Remove the limitation on adding prebuild info only for atoms in the
"partner" category and instead filter the atoms from the prebuild info.

Add "fuchsia.hardware.pty" to test_collection.json to cover the case
where an atom is in another category. The contents of the expected IDK
do not change.

Bug: 417997995
Test: New validation data added for `fx build //build/sdk/generate_prebuild_idk:generate_prebuild_idk_integration_test`
Change-Id: I04e5044c55dc02c0959f15958a06f09cc873c090
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1295091
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>