[firmware] Pass bootloader file from device tree
As fuchsia is moving towards being Android/Linux compatible, we need to
find a standard mechanism for passing bootloader files such as ssh keys.
This CL proposes to use the "/chosen/google/bootloader-files" device
tree node for this purpose, which is being standardized by Android GBL
for passing user defined data from bootloader. A
DevicetreeBootloaderFilesItem class is implemented for scaning and
constructing ZBI items from files in this node.

This will initially be used on moonflower as a necessary step for
eventually converging to Android bootloader.

Bug: 421414311

Change-Id: I586a8332bc24ed79e0860ab039e274e8f3fff37c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1295284
Reviewed-by: David Pursell <dpursell@google.com>
Commit-Queue: Yecheng Zhao <zyecheng@google.com>