[usb_rs] DeviceHandle includes SerialNumber if available
Checks the device identity based on the `stat` of the usb device file
given to the DeviceHandleInner and parses out the special device major
and minor version numbers.

Then scans sysfs for entries that match that major/minor pair.

Then check if there is a `serial` file in the sysfs entry, and if so,
reads the contents of that file to a string.

Getting the device's serial number allows us to migrate off
`usb_bulk` for our Fastboot implementation.

Change-Id: I980bf11bffb75a4d05e2c193d583e1a96a6a636c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1295745
Fuchsia-Auto-Submit: Colin Nelson <colnnelson@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Colin Nelson <colnnelson@google.com>