[build][idk] Ensure no duplicate atom labels in prebuild info
There is no checking whether an entry already exists in `_labels_map` or
`_alias_map` before adding the new atom info. Since we do not expect
this to happen, assert that the combined size matches the number of
atoms in the prebuild info.

Duplicate atom labels could happen if
`sdk_cross_compile_host_tools = true` because the host tools for arm64
and x64 are both built in the main build and not a sub-build. To avoid
this, always append the CPU architecture to the atom label for
"ffx_tool" and "host_tool" atom types.

Bug: 417997995
Change-Id: I71db40ca13d5cc240ec016b53c8318f27381177d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1294739
Fuchsia-Auto-Submit: David Dorwin <ddorwin@google.com>
Reviewed-by: David Turner <digit@google.com>
Commit-Queue: David Dorwin <ddorwin@google.com>