[fdf] [mmio] Add a splittable Mmio implementation that performs operations on memory.
This is intended to be the primary implementation for accessing device
memory. It can also be used with regular memory, e.g. for testing.

Explicit inline assembly is used for the x64 and arm64 implementations,
as the compiler may generate instructions that are incompatible with
KVM.

Change-Id: I307051e6a792417f7570c87194c361a6bed89a36
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1288287
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: Jason Choy <jjwchoy@google.com>
Reviewed-by: David Koloski <dkoloski@google.com>